+++ apps/demo/src/app/docs/markdown/doc/markdown-doc.component.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-markdown-doc',
+  templateUrl: './markdown-doc.component.html',
+  styleUrls: ['./markdown-doc.component.scss'],
+})
+export class MarkdownDocComponent {}