+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Changelog
 
-## 14.1.0-next.0
+## 14.1.0-next.0 (2022-06-19)
 
-* **core:** make use of typed reactive forms
+### Features
+
+* **markdown:** created library by extracting markdown module from core
+
+### Breaking Changes
+
+* **core:** removed markdown module and peer dependency to marked
 
 ## 14.0.1 (2022-06-07)